Added libcurl fetching
