+++ .clang-format
@@ -1,0 +1,3 @@
+BasedOnStyle: llvm
+AllowShortFunctionsOnASingleLine: None
+SortIncludes: false
