+++ .vscode/settings.json
@@ -30,6 +30,8 @@
         "valarray": "c",
         "sbuf.structs.h": "c",
         "cstring": "c",
-        "discord.structs.h": "c"
+        "discord.structs.h": "c",
+        "easy_fetch.c": "c",
+        "easy_fetch.h": "c"
     }
 }
