+++ src/discord-in-c/fetch/Makefile
@@ -1,0 +1,2 @@
+youtube_fetch: youtube_fetch.c
+	gcc -c youtube_fetch.c -lcurl -g -Wall -Wextra -Wno-unused-parameter -fsanitize=address -pedantic
