--- src/discord-in-c/youtube/__tests__/youtube-func.test.c
@@ -1,1 +1,0 @@
-#include "../youtube-func.h"
