--- src/discord-in-c/youtube/youtube-func.test.c
@@ -1,3 +1,0 @@
-
-
-int main(void) {}
